Implement Reconstruction and Misc Fixes
Implement the first version of reconstruction. Work needs to be done to tweak the parameters used here with more testing.

Rename ClipCube to ClipVolume, adding support for non-uniform scaled cubes.

Switch to the KinectV2 SDK and change fusion code related to that.

Hack in a test for the object scan rest endpoint.

Fix commenting style in ServerMain.cpp.

Dramatically change threading and invalidation model of ProgressVisualizer.

Fix camera fitting bugs.